+++ Visor/components/Notch/VisorSkyLightWindow.swift
@@ -6,30 +6,8 @@
 //
 
 import Cocoa
-import SkyLightWindow
 import Defaults
 import Combine
-
-extension SkyLightOperator {
-    func undelegateWindow(_ window: NSWindow) {
-        typealias F_SLSRemoveWindowsFromSpaces = @convention(c) (Int32, CFArray, CFArray) -> Int32
-        
-        let handler = dlopen("/System/Library/PrivateFrameworks/SkyLight.framework/Versions/A/SkyLight", RTLD_NOW)
-        guard let SLSRemoveWindowsFromSpaces = unsafeBitCast(
-            dlsym(handler, "SLSRemoveWindowsFromSpaces"),
-            to: F_SLSRemoveWindowsFromSpaces?.self
-        ) else {
-            return
-        }
-        
-        // Remove the window from the SkyLight space
-        _ = SLSRemoveWindowsFromSpaces(
-            connection,
-            [window.windowNumber] as CFArray,
-            [space] as CFArray
-        )
-    }
-}
 
 class SkyLightWindow: NSPanel {
     private var isSkyLightEnabled: Bool = false